Moved tools to a separate utility file
